+++ R/utils.r
@@ -10,6 +10,10 @@ is_any <- function(obj, klasses) {
   any(vapply(klasses, inherits, logical(1), x = obj))
 }
 
+package_function <- function(pkg, fn) { # for when using :: breaks R CMD check
+  get(fn, envir = getNamespace(pkg))
+}
+
 #' Call a method on an object regardless of its OOP type.
 #'
 #' @name OOP_type_independent_method